+++ packages/gamut-labs/src/experimental/index.ts
@@ -1,8 +1,10 @@
 export * from './Card';
+export * from './Coachmark';
 export * from './Popover';
 export * from './Text';
 export * from './AppHeader';
 export * from './AppHeader/AppHeaderElements/AppHeaderDropdown';
+export * from './AppHeader/AppHeaderElements/types';
 export * from './AppHeader/AppHeaderElements/AppHeaderLink';
 export * from './AppHeader/AppHeaderElements/AppHeaderLogo';
 export * from './AppHeaderMobile';
@@ -10,3 +12,4 @@ export * from './AppHeaderMobile/AppHeaderLinkMobile';
 export * from './AppHeaderMobile/AppHeaderMainMenuMobile';
 export * from './AppHeaderMobile/AppHeaderSubMenuMobile';
 export * from './GlobalHeader';
+export { AnonHeaderVariant } from './GlobalHeader/types';