+++ packages/gamut-styles/src/system.ts
@@ -1,4 +1,5 @@
 import { system } from '@codecademy/gamut-system';
+import isPropValid from '@emotion/is-prop-valid';
 
 import { Theme } from './theme';
 
@@ -70,6 +71,16 @@ const {
   },
 });
 
+const allProps = Object.keys(properties).reduce<string[]>(
+  (carry, prop: keyof typeof properties) => {
+    return [...carry, ...properties[prop].propNames];
+  },
+  []
+);
+
+const shouldForwardProp = (prop: string) =>
+  isPropValid(prop) && !allProps.includes(prop);
+
 export {
   variant,
   properties,
@@ -83,4 +94,5 @@ export {
   shadow,
   space,
   border,
+  shouldForwardProp,
 };