+++ packages/variance/src/index.tsx
@@ -1,0 +1,2 @@
+export { variance } from './core';
+export * from './types/props';