+++ packages/gamut-icons/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-icons@4.0.0...@codecademy/gamut-icons@4.1.0) (2021-03-04)
+
+
+### Features
+
+* LocationPinIcon ([d327380](https://github.com/Codecademy/client-modules/commit/d32738000630d9437a97d570f807626b4f5ec937))
+
+
+
+## [4.0.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-icons@3.15.0...@codecademy/gamut-icons@4.0.0) (2021-02-18)
+
+
+### ⚠ BREAKING CHANGES
+
+* **Banner:** Banner new visual identity [GM-123]
+
+### Features
+
+* **Banner:** Banner new visual identity [GM-123] ([07c26fb](https://github.com/Codecademy/client-modules/commit/07c26fbfcfcc0a3fbb2eba84adcc9dac6645876f))
+
+
+
 ## [3.15.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-icons@3.14.1...@codecademy/gamut-icons@3.15.0) (2021-02-12)