+++ packages/markdown-overrides/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.5.5](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.5.4...@codecademy/markdown-overrides@0.5.5) (2021-03-04)
+
+**Note:** Version bump only for package @codecademy/markdown-overrides
+
+
+
+
+
+### [0.5.4](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.5.3...@codecademy/markdown-overrides@0.5.4) (2021-03-02)
+
+**Note:** Version bump only for package @codecademy/markdown-overrides
+
+
+
+
+
+### [0.5.3](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.5.2...@codecademy/markdown-overrides@0.5.3) (2021-02-23)
+
+**Note:** Version bump only for package @codecademy/markdown-overrides
+
+
+
+
+
 ### [0.5.2](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.5.1...@codecademy/markdown-overrides@0.5.2) (2021-02-17)
 
 **Note:** Version bump only for package @codecademy/markdown-overrides